fix: Pass the github token to all jobs in the workflows.
We weren't passing the GITHUB_TOKEN to both steps of the quarterly task
which was causing it to fail on the first step which didn't pass the
GITHUB_TOKEN.

Rather than just fixing the one step, I updated both the quarterly and
weekly file to pass the GITHUB_TOKEN env to all jobs/steps in their
respective workflow files.  Since these files should be only github
issue creation commands I think this is fine to do and simplifes things
and makes the workflows less error prone.